refactor(keys): give each key-detail view its own file
The dialog held five components. Fingerprint already moved out; the remaining
three views become siblings under components/key-detail/.
