+++ src/components/key-detail/identity-policy.ts
@@ -1,0 +1,16 @@
+import { assertActiveProfile, assertActiveSuite, resolveSuite } from "@/crypto/pq/suites"
+import type { PostQuantumIdentity } from "@/schemas/domain"
+
+export function assertUsableIdentity(identity: PostQuantumIdentity): void {
+  assertActiveProfile(identity.profile)
+  assertActiveSuite(resolveSuite(identity.kem.algorithm, identity.signing.algorithm))
+}
+
+export function isUsableIdentity(identity: PostQuantumIdentity): boolean {
+  try {
+    assertUsableIdentity(identity)
+    return true
+  } catch {
+    return false
+  }
+}
